refactor: mainの分離
